requirements.txt
dnspython3==1.12.0
requests==2.9.1
six==1.10.0
tld==0.7.6
